backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
alembic>=1.13.0
asyncmy>=0.2.9
pymysql>=1.1.0
httpx>=0.26.0
numpy>=2.0.0
scikit-learn>=1.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4
openpyxl>=3.1.0
celery>=5.3.0
redis>=5.0.0
kombu>=5.3.0
